refactor && fix all buttons start loading
